+++ Editor/Core/Components/AssetSkinner/AssetSkinnerGUI.cs
@@ -14,7 +14,6 @@ namespace Miniclip.ShapeShifter.Skinner
     public class AssetSkinnerGUI
     {
         private static Vector2 scrollPosition;
-        private static bool showSkinner = true;
 
         private static readonly string defaultIcon = "WelcomeScreen.AssetStoreLogo";
         private static readonly string errorIcon = "console.erroricon";
@@ -193,7 +192,10 @@ namespace Miniclip.ShapeShifter.Skinner
                         GUILayout.MaxHeight(buttonWidth)
                     );
 
-                    AssetPreviewDropArea(game, path, guid, drawDropAreaToReplace);
+                    if (IsDragAndDroppable(path))
+                    {
+                        AssetPreviewDropArea(game, path, guid, drawDropAreaToReplace);
+                    }
                 }
                 else
                 {
@@ -211,6 +213,11 @@ namespace Miniclip.ShapeShifter.Skinner
             }
         }
 
+        private static bool IsDragAndDroppable(string path)
+        {
+            return IsValidImageFormat(Path.GetExtension(path));
+        }
+
         private static void AssetPreviewDropArea(string game, string path, string guid, bool drawDropAreaToReplace)
         {
             if (drawDropAreaToReplace && DropAreaGUI(out string replacementFilePath))
@@ -218,8 +225,8 @@ namespace Miniclip.ShapeShifter.Skinner
                 if (Path.GetExtension(replacementFilePath) == Path.GetExtension(path)
                     && !PathUtils.IsDirectory(replacementFilePath))
                 {
-                    FileUtil.DeleteFileOrDirectory(path);
-                    FileUtil.CopyFileOrDirectory(replacementFilePath, path);
+                    FileUtils.ValidatePathSafety(path);
+                    FileUtil.ReplaceFile(replacementFilePath, path);
 
                     if (ShapeShifter.ActiveGameName == game)
                     {