+++ Editor/Utils/PathUtils.cs
@@ -9,6 +9,7 @@ namespace Miniclip.ShapeShifter.Utils
     class PathUtils
     {
         private static string ASSETS_FOLDER_NAME => "Assets";
+
         private static string PACKAGES_FOLDER_NAME => "Packages";
 
         internal static bool IsInternalPath(string path)
@@ -141,7 +142,7 @@ namespace Miniclip.ShapeShifter.Utils
         {
             if (string.IsNullOrEmpty(path))
             {
-                throw new ArgumentException("Path is empty");
+                return "";
             }
 
             if (Path.IsPathRooted(path))
@@ -207,7 +208,7 @@ namespace Miniclip.ShapeShifter.Utils
         /// <returns></returns>
         public static bool FileOrDirectoryExists(string assetPath)
         {
-            assetPath = GetFullPath(assetPath);
+            assetPath = NormalizePath(GetFullPath(assetPath));
 
             if (string.IsNullOrEmpty(assetPath))
                 return false;
@@ -223,6 +224,16 @@ namespace Miniclip.ShapeShifter.Utils
             }
 
             return false;
+        }
+
+        public static bool ArePathsEqual(string pathA, string pathB)
+        {
+            return string.Equals(NormalizePath(pathA), NormalizePath(pathB), StringComparison.Ordinal);
+        }
+
+        public static string NormalizePath(string path)
+        {
+            return path.TrimEnd(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar);
         }
     }
 }