+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0]
+
+### Warning
+This version requires adding a "/" to end of any git ignore path related to a folder. 
+
+### Fixed
+- Fixed issue where opening project for the first time would fail to restore any skinned folder
+
+### Added
+- Validating path safety to prevent any possible edge case where a copy/delete operation would target the root folder or any other dangerous target path.
+
 ## [0.3.1]
 ### Fixed
 - Fixed issue where opening skin folders externally would create DS_Store files that fooled shapeshifter